psy_web/requirements.txt
anyascii==0.3.3
asgiref==3.9.1
beautifulsoup4==4.13.5
black==25.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
defusedxml==0.7.1
Django==5.2.5
django-filter==25.1
django-modelcluster==6.4
django-permissionedforms==0.1
django-stubs-ext==5.2.2
django-taggit==6.1.0
django-tasks==0.8.1
django-treebeard==4.7.1
djangorestframework==3.16.1
draftjs_exporter==5.1.0
et_xmlfile==2.0.0
filetype==1.2.0
gunicorn==23.0.0
idna==3.10
laces==0.1.2
mypy_extensions==1.1.0
openpyxl==3.1.5
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
pillow_heif==1.1.0
platformdirs==4.4.0
psycopg2-binary==2.9.10
pytokens==0.1.10
requests==2.32.5
ruff==0.13.2
sentry-sdk==2.37.1
soupsieve==2.7
sqlparse==0.5.3
telepath==0.3.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wagtail==7.1
Willow==1.11.0
